main-application/pyproject.toml
[project]
name = "atmos-fractal"
version = "0.1.0"
description = "Weather-based fractal and audio observer"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "requests>=2.32.0",
    "numpy>=1.26.0",
    "numba>=0.60.0",
    "scipy>=1.13.0",
    "pillow>=10.4.0",
    "matplotlib>=3.10.8",
    "jinja2>=3.1.6",
]
